Migrate updateNodeAtIndex and friends
