+++ packages/frontend/editor-ui/src/types/canvas.ts
@@ -7,11 +7,12 @@ import type {
 	OnConnectStartParams,
 	ViewportTransform,
 } from '@vue-flow/core';
-import type { IExecutionResponse, INodeUi } from '@/Interface';
+import type { IExecutionResponse, INodeUi, INodeUpdatePropertiesInformation } from '@/Interface';
 import type { ComputedRef, Ref } from 'vue';
 import type { EventBus } from '@n8n/utils/event-bus';
 import type { CanvasLayoutSource } from '@/composables/useCanvasLayout';
 import type { NodeIconSource } from '../utils/nodeIcon';
+import type { WorkflowState } from '@/composables/useWorkflowState';
 
 export const enum CanvasConnectionMode {
 	Input = 'inputs',
@@ -187,6 +188,10 @@ export type CanvasEventBusEvents = {
 	};
 	tidyUp: { source: CanvasLayoutSource; nodeIdsFilter?: string[]; trackEvents?: boolean };
 	'create:sticky': never;
+};
+
+export type WorkflowStateBusEvents = {
+	updateNodeProperties: [WorkflowState, INodeUpdatePropertiesInformation];
 };
 
 export interface CanvasNodeInjectionData {
